refactor(core): run tests through the public createRstest API
Route non-watch CLI runs through the instance API. Keep process-level
behavior in a private runtime channel: embedded mode, tracing, package
installation confirmation, and host exit-code mirroring stay out of
CreateRstestOptions. Preserve the existing watch and list drivers.

Expose every run flag through RunOptions except creation-time config,
configLoader and root, plus CLI-only trace. Derive CommonOptions from
RunOptions and resolve the findRelatedTests alias at the parse boundary.

Preserve exclude.override and clone config arrays so operation overrides
do not mutate the instance configuration before defaults are applied.
