+++ e2e/programmatic/fixtures/cli-missing-deps/index.test.js
@@ -1,0 +1,1 @@
+test('works', () => {});
