+++ e2e/projects/fixtures-duplicate-names/alpha/rstest.config.mjs
@@ -1,0 +1,1 @@
+export default { name: 'alpha' };
