+++ e2e/projects/fixtures-duplicate-names/b1/rstest.config.mjs
@@ -1,0 +1,1 @@
+export default { name: 'beta' };
