+++ e2e/projects/fixtures-duplicate-names/rstest.config.mjs
@@ -1,0 +1,3 @@
+export default {
+  projects: ['alpha', 'b1', 'b2'],
+};
