+++ e2e/projects/index.test.ts
@@ -7,7 +7,45 @@ const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
 
 describe('test projects', () => {
+  it('validates duplicate names only among selected projects', async () => {
+    const options = {
+      nodeOptions: { cwd: join(__dirname, 'fixtures-duplicate-names') },
+    };
+    const selected = await runRstestCli({
+      command: 'rstest',
+      args: ['run', '--project', 'alpha'],
+      options,
+    });
+    await selected.expectExecSuccess();
+    expect(selected.cli.stdout).toContain('Tests 1 passed');
+
+    const all = await runRstestCli({
+      command: 'rstest',
+      args: ['run'],
+      options,
+    });
+    await all.expectExecFailed();
+    all.expectStderrLog('Project name "beta" is already used');
+  });
+
   describe('merge configs', () => {
+    it('replays exclusions onto every file-based project', async () => {
+      const { cli, expectExecSuccess } = await runRstestCli({
+        command: 'rstest',
+        args: [
+          'run',
+          '--globals',
+          '--exclude',
+          '**/*.test.*',
+          '--passWithNoTests',
+        ],
+        options: { nodeOptions: { cwd: join(__dirname, 'fixtures') } },
+      });
+      await expectExecSuccess();
+      expect(cli.stdout).toContain('No test files found');
+      expect(cli.stdout).toContain('Test Files no tests');
+    });
+
     it('should run projects correctly with cli options', async () => {
       const { cli, expectExecSuccess, expectLog } = await runRstestCli({
         command: 'rstest',
