+++ packages/core/src/cli/init.ts
@@ -2,14 +2,9 @@ import { existsSync, readFileSync, statSync } from 'node:fs';
 import type { LoadConfigOptions } from '@rsbuild/core';
 import { basename, dirname, resolve } from 'pathe';
 import { type GlobOptions, glob, isDynamicPattern } from 'tinyglobby';
+import type { RunOptions } from '../api/types';
 import { loadConfig, plainDeepMerge, resolveExtends } from '../config';
-import type {
-  BrowserName,
-  Project,
-  ResolvedRstestConfig,
-  RstestConfig,
-  RstestOutputConfig,
-} from '../types';
+import type { Project, ResolvedRstestConfig, RstestConfig } from '../types';
 import {
   castArray,
   color,
@@ -20,100 +15,26 @@ import {
   logger,
 } from '../utils';
 
-export type CommonOptions = {
+export type CommonOptions = Omit<
+  RunOptions,
+  'filters' | 'filterMode' | 'shard'
+> & {
   root?: string;
   config?: string;
   configLoader?: LoadConfigOptions['loader'];
-  related?: boolean;
-  findRelatedTests?: boolean;
-  changed?: boolean | string;
-  globals?: boolean;
-  /**
-   * Pool options.
-   * - `string`: shorthand for `{ type: string }` (from `--pool` flag)
-   * - `object`: detailed pool config (from `--pool.*` options)
-   */
-  pool?:
-    | string
-    | {
-        type?: string;
-        maxWorkers?: string | number;
-        execArgv?: string[] | string;
-      };
-  /**
-   * Browser mode options.
-   * - `boolean`: shorthand for `{ enabled: boolean }` (from `--browser` flag)
-   * - `object`: detailed browser config (from `--browser.*` options)
-   */
-  browser?:
-    | boolean
-    | {
-        enabled?: boolean;
-        name?: BrowserName;
-        headless?: boolean;
-        port?: number;
-        strictPort?: boolean;
-        providerOptions?: Record<string, unknown>;
-      };
-  isolate?: boolean;
-  include?: string[];
-  exclude?: string[];
-  reporters?: string | string[];
-  project?: string[];
-  /**
-   * Coverage options.
-   * - `boolean`: shorthand for `{ enabled: boolean }` (from `--coverage` flag)
-   * - `object`: detailed coverage config (from `--coverage.*` options)
-   */
-  coverage?:
-    | boolean
-    | {
-        enabled?: boolean | string;
-        allowExternal?: boolean;
-        provider?: 'istanbul' | 'v8';
-        include?: string | string[];
-        changed?: boolean | string;
-        exclude?: string | string[];
-        reporters?: string | string[];
-        reportsDirectory?: string;
-        reportOnFailure?: boolean | string;
-        clean?: boolean | string;
-      };
-  passWithNoTests?: boolean;
-  onlyFailures?: boolean;
-  silent?: boolean | 'passed-only';
-  printConsoleTrace?: boolean;
-  logHeapUsage?: boolean;
-  detectAsyncLeaks?: boolean;
   trace?: boolean;
-  disableConsoleIntercept?: boolean;
-  update?: boolean;
-  testNamePattern?: RegExp | string;
-  testTimeout?: number;
-  hookTimeout?: number;
-  testEnvironment?: string;
-  clearMocks?: boolean;
-  resetMocks?: boolean;
-  restoreMocks?: boolean;
-  unstubGlobals?: boolean;
-  unstubEnvs?: boolean;
-  retry?: number;
-  maxConcurrency?: number;
-  slowTestThreshold?: number;
-  hideSkippedTests?: boolean;
-  hideSkippedTestFiles?: boolean;
-  bail?: number | boolean;
   shard?: string;
-  includeTaskLocation?: boolean;
-  federation?: boolean;
-  source?: {
-    tsconfigPath?: string;
-  };
-  dev?: {
-    writeToDisk?: boolean;
-  };
-  output?: Pick<RstestOutputConfig, 'emitAssets' | 'cleanDistPath' | 'module'>;
 };
+
+export const loadCliConfig = (
+  options: CommonOptions,
+  cwd: string,
+): ReturnType<typeof loadConfig> =>
+  loadConfig({
+    cwd: options.root ? getAbsolutePath(cwd, options.root) : cwd,
+    path: options.config,
+    configLoader: options.configLoader,
+  });
 
 function coerceCliBoolean(value: unknown): boolean | undefined {
   if (typeof value === 'boolean') {
@@ -605,22 +526,6 @@ export async function resolveProjects({
   if (!projects.length) {
     throw formatNoProjectsFoundError(config, options.project);
   }
-
-  const names = new Set<string>();
-
-  projects.forEach((project) => {
-    if (names.has(project.config.name!)) {
-      const conflictProjects = projects.filter(
-        (p) => p.config.name === project.config.name,
-      );
-      throw `Project name "${project.config.name}" is already used. Please ensure all projects have unique names.
-Conflicting projects:
-${conflictProjects.map((p) => `- ${p.configFilePath || p.config.root}`).join('\n')}
-        `;
-    }
-
-    names.add(project.config.name!);
-  });
 
   return projects;
 }
